add changes from #389
