+++ nemo_curator/scripts/semdedup/compute_embeddings.py
@@ -16,10 +16,10 @@ import logging
 import os
 import time
 
+from nemo_curator import EmbeddingCreator
 from nemo_curator.datasets import DocumentDataset
 from nemo_curator.log import create_logger
 from nemo_curator.modules.config import SemDedupConfig
-from nemo_curator.modules.semantic_dedup import EmbeddingCreator
 from nemo_curator.utils.distributed_utils import get_client, read_data
 from nemo_curator.utils.file_utils import expand_outdir_and_mkdir, get_remaining_files
 from nemo_curator.utils.script_utils import ArgumentHelper
