+++ tutorials/zyda2-tutorial/1_fuzzy_dedup/2_buckets_to_edges.py
@@ -4,8 +4,8 @@ import time
 
 import dask_cudf
 
+from nemo_curator import BucketsToEdges
 from nemo_curator.datasets import DocumentDataset
-from nemo_curator.modules.fuzzy_dedup import BucketsToEdges
 from nemo_curator.utils.distributed_utils import get_client, get_num_workers
 
 logging.basicConfig(format="%(asctime)s: %(message)s", level=logging.INFO)
